- Moved reader files into inputReader directory and added inteface - implemented deletion of old output files befor running the program
